Refactor: 뷰 분리
