+++ Projects/CodePlay/CodePlay/Presentation/Scene/ExportPlaylist/View/ExportPlaylistView.swift
@@ -54,61 +54,3 @@ struct ExportPlaylistView: View {
     }
 }
 
-// MARK: 애플뮤직 플레이리스트로 전송하는 뷰 (hifi 06_1부분)
-struct ExportLoadingView: View {
-    @State private var progress: Double = 0.0
-
-    var body: some View {
-        VStack(spacing: 24) {
-            Spacer()
-
-            Text("Apple Music으로 전송 중...")
-                .font(.title3)
-
-            ProgressView(value: progress)
-                .progressViewStyle(.linear)
-                .padding(.horizontal, 32)
-
-            Spacer()
-        }
-        .onAppear {
-            withAnimation(.easeInOut(duration: 5)) {
-                progress = 1.0
-            }
-        }
-    }
-}
-
-// MARK: 전송 완료 이후, 애플뮤직 앱으로 전환하는 뷰 (hifi 07_1부분)
-struct ExportSuccessView: View {
-    @Environment(\.dismiss) private var dismiss
-
-    var body: some View {
-        NavigationStack {
-            VStack(spacing: 32) {
-                Spacer()
-
-                Text("🎉 전송이 완료되었습니다!")
-                    .font(.title2)
-                    .multilineTextAlignment(.center)
-
-                BottomButton(title: "Apple Music으로 이동") {
-                    if let url = URL(string: "music://") {
-                        UIApplication.shared.open(url)
-                    }
-                }
-                .padding(.horizontal, 16)
-
-                Spacer()
-            }
-            .navigationTitle("전송 완료")
-            .toolbar {
-                ToolbarItem(placement: .navigationBarLeading) {
-                    Button("닫기") {
-                        dismiss()
-                    }
-                }
-            }
-        }
-    }
-}
